@@ -1,22 +1,29 @@
 package com.github.willgutierrez.example
 
-import android.arch.lifecycle.ViewModelProviders
-import android.support.v7.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProviders
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import com.github.willgutierrez.example.funtionalprograming.Game
 import com.github.willgutierrez.example.funtionalprograming.UtilsFunctionalPrograming
+import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.item_user_layout.view.*
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mainViewModel: MainViewModel
+    private val adapter = listAdapter()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        setupRecyclerView()
         mainViewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
-        mainViewModel.itemList.observe(this, android.arch.lifecycle.Observer { items ->
-            items?.let { showText(it.joinToString { it }) }
+        mainViewModel.itemList.observe(this, { items ->
+            items?.let { result ->
+                showText(result.joinToString { it })
+                adapter.submitList(result.toItemList())
+            }
         })
         mainViewModel.getFirstStudents()
 
@@ -24,6 +31,10 @@ class MainActivity : AppCompatActivity() {
 
         Game().simulate()
 
+    }
+
+    private fun setupRecyclerView() {
+        recyclerView.adapter = adapter
     }
 
     private fun printPairVals() {
@@ -34,4 +45,19 @@ class MainActivity : AppCompatActivity() {
 
     private fun showText(text: String) =
             Log.d("logMain", text)
+}
+
+private fun List<String>.toItemList(): List<Item> =
+        map {
+            UserItem(it)
+        }
+
+class UserItem(override val content: String) : Item {
+    override val layoutResource: Int = R.layout.item_user_layout
+    override val id = javaClass
+
+    override fun AdapterLayout.bind() {
+        userName.text = content
+    }
+
 }
